// spike/api-ergonomics/src/types.ts
export interface RunResult {
  conclusion: 'success' | 'failure';
  outputs: Record<string, string>;
  steps: StepResult[];
  step(id: string): StepResult | undefined;
  env: Record<string, string>;
  annotations: Annotation[];
  readonly stdout: string;
  readonly stderr: string;
}

export interface StepResult {
  id: string;
  name: string;
  phase: 'pre' | 'main' | 'post';
  ran: boolean;
  outcome: 'success' | 'failure' | 'skipped';
  conclusion: 'success' | 'failure' | 'skipped';
  outputs: Record<string, string>;
  if?: { expression: string; result: boolean };
  uses?: { ref: string; mocked: boolean };
  stdout: string;
  stderr: string;
}

export interface Annotation {
  level: 'error' | 'warning' | 'notice' | 'debug';
  message: string;
  file?: string;
  line?: number;
  col?: number;
}

export interface RunInput {
  inputs?: Record<string, string | number | boolean>;
  env?: Record<string, string>;
  github?: Partial<GitHubContext>;
  runner?: Partial<RunnerContext>;
  secrets?: Record<string, string>;
  jobStatus?: 'success' | 'failure' | 'cancelled';
}

export interface GitHubContext {
  repository: string;
  repositoryOwner: string;
  sha: string;
  ref: string;
  refName: string;
  refType: string;
  actor: string;
  eventName: string;
  workflow: string;
  runId: string;
  runNumber: string;
  runAttempt: string;
  serverUrl: string;
  apiUrl: string;
  [key: string]: unknown;
}

export interface RunnerContext {
  os: string;
  arch: string;
  name: string;
  temp: string;
  toolCache: string;
}
// ...
export interface ParsedInput {
  description?: string;
  required?: boolean;
  default?: string;
}

export interface ParsedOutput {
  description?: string;
  value?: string;
}

export interface ParsedStep {
  id?: string;
  name?: string;
  if?: string;
  uses?: string;
  with?: Record<string, string>;
  run?: string;
  shell?: string;
  env?: Record<string, string>;
  'working-directory'?: string;
  'continue-on-error'?: boolean | string;
  'timeout-minutes'?: number;
}

export interface ParsedAction {
  name: string;
  description?: string;
  inputs?: Record<string, ParsedInput>;
  outputs?: Record<string, ParsedOutput>;
  runs: {
    using: string;
    steps?: ParsedStep[];
    // node
    pre?: string;
    'pre-if'?: string;
    main?: string;
    post?: string;
    'post-if'?: string;
  };
}